fix: move download logic to service
