+++ projects/components/src/table/cells/csv-generators/table-cell-boolean-csv-generator.ts
@@ -4,7 +4,6 @@ import { isNil } from 'lodash-es';
 
 @Injectable({ providedIn: 'root' })
 export class TableCellBooleanCsvGenerator implements TableCellCsvGenerator<boolean> {
-  public readonly type = 'CSV_GENERATOR';
   public readonly cellType: string = 'boolean';
 
   public generateSafeCsv(cellData?: boolean): string | undefined {
