+++ projects/components/src/table/cells/csv-generators/table-cell-string-csv-generator.ts
@@ -4,7 +4,6 @@ import { CoreTableCellRendererType } from '../types/core-table-cell-renderer-typ
 
 @Injectable({ providedIn: 'root' })
 export class TableCellStringCsvGenerator implements TableCellCsvGenerator<string> {
-  public readonly type = 'CSV_GENERATOR';
   public cellType: string = CoreTableCellRendererType.Text;
 
   public generateSafeCsv(cellData?: string): string | undefined {
